+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexMacroDeclarationImpl.java
@@ -1,8 +1,4 @@
 /*
- * ----------------------------------------------------------------
- * --- WARNING: THIS FILE IS GENERATED AND WILL BE OVERWRITTEN! ---
- * ----------------------------------------------------------------
- *
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
  *
@@ -63,6 +59,12 @@ public class ImpexMacroDeclarationImpl extends ASTWrapperPsiElement implements I
   @NotNull
   public List<ImpexMacroValueDec> getMacroValueDecList() {
     return PsiTreeUtil.getChildrenOfTypeAsList(this, ImpexMacroValueDec.class);
+  }
+
+  @Override
+  @NotNull
+  public List<ImpexString> getStringList() {
+    return PsiTreeUtil.getChildrenOfTypeAsList(this, ImpexString.class);
   }
 
 }